Time: 66 ms (30.19%), Space: 90 MB (25.76%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.cpp
@@ -14,7 +14,7 @@ public:
         if(n%2==0)
         {
             
-            return (double)((double)(v[n/2]+v[(n/2)-1])/2);
+            return ((double)(v[n/2]+v[(n/2)-1])/2);
         }
         else
         {
